Compatibility with Wagtail 5.x #22 (#23)
* Update compatibility and exception handling

Introduce backward compatibility for import statements to accommodate changes in different versions of Wagtail. The import statements from Wagtail modules now have exception handling to ensure compatibility across multiple versions.

Code has also been updated to use double quotes for strings consistently across the entire module.

Additionally, the project version was updated to 0.9.2a and a minor adjustment was made to the setup file to reflect the changes in the Wagtail compatibility range.

These changes aim to improve the module's robustness and future-proof its code to handle possible deprecations or changes.

* Update gallery_tags initialization in wagtail gallery model

Replaced mutable default argument in 'get_gallery_tags' method with a non-mutable default. This change prevents potential bugs related to Python's mutable default arguments. Simplified the 'tag_archive' method by removing exception handling for appending the tag to the 'taglist'. This simplification is reliable because 'taglist' is always initialized as an empty list.

* Adjust wagtail import for compatibility in wagtail_simple_gallery

Adding exception handling for the import of the wagtail model in wagtail_simple_gallery in order to maintain compatibility with different versions of wagtail. This import has been adjusted throughout all the edited files providing a fail-safe for cases when the path to import wagtail's core changes.

---------

Co-authored-by: François GUÉRIN <fguerin@ville-tourcoing.fr>

Thanks!